Conflicts:
	recipes/wlan/qcacld30-ll_git.bb

+++ recipes/wlan/wcnss_git.bb
@@ -3,11 +3,13 @@ LIC_FILES_CHKSUM = "file://${COREBASE}/meta/files/common-licenses/BSD;md5=377548
 LICENSE = "BSD"
 
 PR = "r1"
-
 FILESPATH =+ "${WORKSPACE}:"
 SRC_URI = "file://qcom-opensource/wlan/prima/firmware_bin \
            file://set_wcnss_mode"
 SRC_URI += "file://wcnss_wlan.service"
+SRC_URI += "file://device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_cfg.ini"
+SRC_URI += "file://device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_wlan_nv.bin"
+SRC_URI += "file://device/qcom/wlan/${SOC_FAMILY}/WCNSS_wlan_dictionary.dat"
 
 S = "${WORKDIR}/qcom-opensource/wlan/firmware_bin"
 
@@ -35,8 +37,20 @@ do_install() {
                install "${WORKDIR}"/set_wcnss_mode ${D}/etc/init.d
 	fi
 
-    mkdir -p ${D}/lib/firmware/wlan/prima
-    cp -pP ${WORKSPACE}/android_compat/device/qcom/${SOC_FAMILY}/WCNSS_qcom_cfg.ini ${D}/lib/firmware/wlan/prima
+    install -d ${D}/lib/firmware/wlan/prima/
+    install -d ${D}/persist/
+
+    if [ -e "${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_cfg.ini" ];then
+	    install -m 0644 ${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_cfg.ini ${D}/lib/firmware/wlan/prima/WCNSS_qcom_cfg.ini
+    elif [ -e "${WORKSPACE}/android_compat/device/qcom/${SOC_FAMILY}/WCNSS_qcom_cfg.ini" ];then
+	    cp -pP ${WORKSPACE}/android_compat/device/qcom/${SOC_FAMILY}/WCNSS_qcom_cfg.ini ${D}/lib/firmware/wlan/prima
+    fi
+    if [ -e "${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_wlan_nv.bin" ];then
+	    install -m 0644 "${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_qcom_wlan_nv.bin" ${D}/persist/WCNSS_qcom_wlan_nv.bin
+    fi
+    if [ -e "${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_wlan_dictionary.dat" ]; then
+	    install -m 0644 "${WORKDIR}/device/qcom/wlan/${SOC_FAMILY}/WCNSS_wlan_dictionary.dat" ${D}/persist/WCNSS_wlan_dictionary.dat
+    fi
 }
 do_install_append() {
    install -d ${D}/lib/firmware/wlan/prima
@@ -49,4 +63,5 @@ INITSCRIPT_PARAMS = "start 60 2 3 4 5 . stop 20 0 1 6 ."
 FILES_${PN} = "/lib/firmware/*"
 FILES_${PN} += "/etc/*"
 FILES_${PN} += "/lib/firmware/wlan/prima/*"
+FILES_${PN} += "/persist/*"
